requirements.txt
amqp==5.0.9
asgiref==3.4.1
astroid==2.9.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.11
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-pymemcache==1.0.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
flake8==4.0.1
idna==3.3
isort==5.10.1
kombu==5.2.3
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.1
prompt-toolkit==3.0.24
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pymemcache==3.5.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
